refactor: extract importBoardFromUrl into its own module
board-sets-store now holds only external-store + DB CRUD; the
network-fetch utility lives in its own file.
