+++ src/features/board/storage/board-sets-store.ts
@@ -88,25 +88,6 @@ export async function importBoardFiles(
   return results;
 }
 
-export async function importBoardFromUrl(url: string): Promise<ImportResult> {
-  const response = await fetch(url);
-
-  if (!response.ok) {
-    throw new Error(`Failed to fetch board: HTTP ${response.status}`);
-  }
-
-  const blob = await response.blob();
-  const pathname = new URL(url, window.location.origin).pathname;
-  const filename = pathname.split("/").at(-1) ?? "board.obz";
-
-  const file = new File([blob], filename, {
-    type: blob.type || "application/octet-stream",
-  });
-
-  const [result] = await importBoardFiles(file);
-  return result;
-}
-
 export async function removeBoardSet(setId: string): Promise<void> {
   await withBoardsDB((db) => deleteBoardSet(db, setId));
   await reloadAndBroadcast();
